backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
httpx
sqlalchemy
asyncpg
greenlet
python-multipart
Pillow
edge-tts
yfinance
pandas
cachetools
requests
